[WIP] Started to refactor the ImageTool API
Hopefully will be able to remove (rework) the MediaHelper utility class.
