+++ src/assembly/Our.Umbraco.EditorJs.csproj
@@ -107,8 +107,6 @@
     <Compile Include="Controllers\LinkToolController.cs" />
     <Compile Include="DataEditors\EditorJsConfigurationEditor.cs" />
     <Compile Include="DataEditors\EditorJsDataEditor.cs" />
-    <Compile Include="Helpers\MediaHelper.cs" />
-    <Compile Include="Models\UploadResponse.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ValueConverters\EditorJsModel.cs" />
     <Compile Include="ValueConverters\EditorJsValueConverter.cs" />
